Radio -> Multiple choice
Our Radio component is most of the way to a Checkbox.
Since we're sharing most of the markup and all of the
same CSS, I figured I could stick them both in the same
file.
Called it 'multiple choice' because that's what we called
them at GDS when I worked there and apparently they still
call them that.