Add custom Checkbox styling
Just like we made some changes to the GOVUK radios, we
want to make similar changes to our checkboxes so that
they look like they belong on the same website.

Basically this means making them smaller and fiddling
with the spacing.

It's all magic numbers, but it can't be helped. 😣

Also, interestingly, it seems like we can pass custom
components into redux-form rather than make our custom
components into <Field>s.
https://redux-form.com/6.0.0-alpha.4/docs/faq/customcomponent.md/
Didn't do it in this instance, but useful to know for
next time.