Unique checkbox "name" values even though it's wrong
Like radio buttons, the right way to do checkboxes when
they are all related options is to:
- use the same "name" for all of them
- use different "value"s and "id"s

If the name was "energySource", and values were "oil" and
"electricity", the URL params for the submitted form would
look like `?energySource=oil&energySource=electricity`.

As documented on MDN:
- https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input/checkbox#Examples

***

For some reason, Redux forms breaks this behaviour.
- all "value"s are overwritten to either "true" or "false"
- all checkboxes with identical "name"s are selected
  as a group

I found a few links to this behaviour.

- https://www.reddit.com/r/reactjs/comments/4enekc/redux_form_question/
- https://github.com/erikras/redux-form/issues/2922
- https://github.com/erikras/redux-form/issues/1993

In order to get the right way of doing things to work,
we'd have to manage the behaviour ourselves, which is
not something I am interested in.
Changing the "name"s of our checkboxes so that we can
just move forwards with this.