Ask for both ersRating and eghRating & display one
Now that most of the data has been added, the majority of it only has
the old 0/100 rating (aka eghRating) not the new gigajoule rating (aka
ersRating). This change data broke some of our assumptions about what was
going to be available so we ended up displaying nothing in the majority
of cases.