WIP: deploy with Kubernetes
What's going on here is figuring out what deploying to Kubernetes should look like.
The appeal of k8s is that it limits vendor lock-in with cloud providers while
providing a way to meet user expectations for uptime and organizational
expectations of visibility by managing our container deployments.

Specifically what is being explored here is the developer work flow using the
tools skaffold and kustomize.